+++ client/src/api.ts
@@ -17,8 +17,8 @@ export const apis = {
     const { data } = await base.get('/notion', { params: { id } })
     return data
   },
-  getPage: async (url: string) => {
-    const { data } = await base.get('/page', { params: { url } })
+  getSubstackPage: async (url: string) => {
+    const { data } = await base.get('/substack', { params: { url } })
     return data
   },
   getSameSitePageIds: async (id: string, depth = 0): Promise<string[]> => {
